refactor: 🎨 Big code refactoring
Code refactoring, folder and file structure change to reflect application architecture.
