+++ ormconfig.json
@@ -6,13 +6,13 @@
   "password": "challenge",
   "database": "ton_api",
   "migrations": [
-    "./src/database/migrations/*.ts"
+    "./src/shared/infra/typeorm/migrations/*.ts"
   ],
   "entities": [
-    "./src/modules/account/entities/*.ts",
-    "./src/modules/hits/entities/*.ts"
+    "./src/modules/account/infra/typeorm/entities/*.ts",
+    "./src/modules/hits/infra/typeorm/entities/*.ts"
   ],
   "cli": {
-    "migrationsDir": "./src/database/migrations"
+    "migrationsDir": "./src/shared/infra/typeorm/migrations"
   }
 }
