--- src/config/appConfig.ts
@@ -1,4 +1,0 @@
-
-const appKey = process.env.APP_KEY;
-
-export { appKey };
