+++ src/modules/account/useCases/createUser/CreateUserUseCase.ts
@@ -1,7 +1,7 @@
 import { inject, injectable } from "tsyringe";
-import { AppError } from "../../../../errors/AppError";
 
-import { IUsersRepository } from "../../repositories/contracts/IUsersRepository";
+import { AppError } from "@shared/errors/AppError";
+import { IUsersRepository } from "@modules/account/repositories/IUsersRepository";
 
 interface IRequest {
   name: string;
