+++ src/modules/account/useCases/listUsers/ListUsersController.ts
@@ -1,5 +1,6 @@
 import { Request, Response } from "express";
 import { container } from "tsyringe";
+
 import { ListUsersUseCase } from "./ListUsersUseCase";
