+++ src/modules/account/useCases/updateUser/UpdateUserController.ts
@@ -1,6 +1,7 @@
 import { Request, Response } from "express";
 import { container } from "tsyringe";
-import { AppError } from "../../../../errors/AppError";
+
+import { AppError } from "@shared/errors/AppError";
 import { UpdateUserUseCase } from "./UpdateUserUseCase";
