+++ src/modules/account/useCases/updateUser/UpdateUserUseCase.ts
@@ -1,8 +1,8 @@
 import { hash } from "bcrypt";
 import { inject, injectable } from "tsyringe";
-import { AppError } from "../../../../errors/AppError";
 
-import { IUsersRepository } from "../../repositories/contracts/IUsersRepository";
+import { AppError } from "@shared/errors/AppError";
+import { IUsersRepository } from "@modules/account/repositories/IUsersRepository";
 
 interface IRequest {
   id: string;
