+++ src/modules/hits/useCases/viewCounter/ViewCounterUseCase.ts
@@ -1,8 +1,9 @@
 import { inject, injectable } from "tsyringe";
-import { AppError } from "../../../../errors/AppError";
-import { ICountersRepository } from "../../repositories/contracts/ICountersRepository";
-import { Counter } from "../../entities/Counter";
-import { countApi } from "../../../../services/CountApi";
+
+import { AppError } from "@shared/errors/AppError";
+import { Counter } from "@modules/hits/infra/typeorm/entities/Counter";
+import { ICountersRepository } from "@modules/hits/repositories/ICountersRepository";
+import { countApi } from "@shared/infra/services/CountApi";
 
 interface IViewCounter {
   id: string;
