+++ src/shared/container/index.ts
@@ -1,9 +1,10 @@
 import { container } from "tsyringe";
 
-import { IUsersRepository } from "../../modules/account/repositories/contracts/IUsersRepository";
-import { UsersRepository } from "../../modules/account/repositories/UsersRepository";
-import { ICountersRepository } from "../../modules/hits/repositories/contracts/ICountersRepository";
-import { CountersRepository } from "../../modules/hits/repositories/CountersRepository";
+import { UsersRepository } from "@modules/account/infra/typeorm/repositories/UsersRepository";
+import { IUsersRepository } from "@modules/account/repositories/IUsersRepository";
+import { CountersRepository } from "@modules/hits/infra/typeorm/repositories/CountersRepository";
+import { ICountersRepository } from "@modules/hits/repositories/ICountersRepository";
+
 
 container.registerSingleton<IUsersRepository>(
   "UsersRepository", UsersRepository
