+++ src/shared/infra/http/middlewares/ensureAuthenticated.ts
@@ -1,7 +1,8 @@
 import { NextFunction, Request, Response } from "express";
 import { verify } from "jsonwebtoken";
-import { AppError } from "../errors/AppError";
-import { UsersRepository } from "../modules/account/repositories/UsersRepository";
+
+import { AppError } from "@shared/errors/AppError";
+import { UsersRepository } from "@modules/account/infra/typeorm/repositories/UsersRepository";
 
 interface ITokenPayload {
   sub: string
