+++ src/shared/infra/http/routes/authenticate.routes.ts
@@ -1,6 +1,7 @@
 import { Router } from "express";
 
-import { AuthenticateUserController } from "../modules/account/useCases/authenticateUser/AuthenticateUserController";
+import { AuthenticateUserController } from "@modules/account/useCases/authenticateUser/AuthenticateUserController";
+
 
 const authenticateRoutes = Router();
