+++ src/shared/infra/http/routes/users.routes.ts
@@ -1,9 +1,9 @@
 import { Router } from "express";
 
-import { ListUsersController } from "../modules/account/useCases/listUsers/ListUsersController";
-import { CreateUserController } from "../modules/account/useCases/createUser/CreateUserController";
-import { UpdateUserController } from "../modules/account/useCases/updateUser/UpdateUserController";
-import { ensureAuthenticated } from "../middlewares/ensureAuthenticated";
+import { ListUsersController } from "@modules/account/useCases/listUsers/ListUsersController";
+import { CreateUserController } from "@modules/account/useCases/createUser/CreateUserController";
+import { UpdateUserController } from "@modules/account/useCases/updateUser/UpdateUserController";
+import { ensureAuthenticated } from "@shared/infra/http/middlewares/ensureAuthenticated";
 
 const usersRoutes = Router();
