+++ src/shared/infra/http/server.ts
@@ -2,12 +2,12 @@ import express, { NextFunction, Request, Response } from "express";
 import "express-async-errors";
 import swaggerUi from "swagger-ui-express";
 
-import "./database";
-import "./shared/container";
+import "@shared/infra/typeorm";
+import "@shared/container";
 
-import { router } from "./routes";
-import swaggerFile from "./swagger.json";
-import { AppError } from "./errors/AppError";
+import { router } from "@shared/infra/http/routes";
+import swaggerFile from "../../../swagger.json";
+import { AppError } from "@shared/errors/AppError";
 
 
 const app = express();
